Reverted dropbox change
Dropbox tray icon still doesn't work right on kde as it endlessly attempts to connect. On gnome it syncs, but tray icon is almost completely broken. Takes numerous right clicks to display the context menu and when it does it doesnt say its connected and everything is broken except opening settings and exiting.